refactor: move pgp tests to submodule
